d3d: Share the bounding box code.
